+++ tests/testthat.R
@@ -1,4 +1,8 @@
 library(testthat)
 library(epipredict)
+library(parsnip)
+library(workflows)
+library(dplyr)
+
 
 test_check("epipredict")