+++ tests/testthat/test-epi_workflow.R
@@ -59,7 +59,7 @@ test_that("model can be added/updated/removed from epi_workflow", {
   expect_equal(class(model_spec2), c("linear_reg", "model_spec"))
 
   wf <- remove_model(wf)
-  expect_error(extract_spec_parsnip(wf))
+  expect_snapshot(error = TRUE, extract_spec_parsnip(wf))
   expect_equal(wf$fit$actions$model$spec, NULL)
 })
 
@@ -79,17 +79,11 @@ test_that("forecast method works", {
     ))
   )
 
-  args <- list(
-    fill_locf = TRUE,
-    n_recent = 360 * 3,
-    forecast_date = as.Date("2024-01-01")
-  )
   expect_equal(
-    forecast(wf, !!!args),
+    forecast(wf),
     predict(wf, new_data = get_test_data(
       hardhat::extract_preprocessor(wf),
-      jhu,
-      !!!args
+      jhu
     ))
   )
 })
@@ -103,5 +97,42 @@ test_that("forecast method errors when workflow not fit", {
     step_epi_naomit()
   wf <- epi_workflow(r, parsnip::linear_reg())
 
-  expect_error(forecast(wf))
+  expect_snapshot(error = TRUE, forecast(wf))
+})
+
+test_that("fit method does not silently drop the class", {
+  # This is issue #363
+
+  library(recipes)
+  tbl <- tibble::tibble(
+    geo_value = 1,
+    time_value = 1:100,
+    x = 1:100,
+    y = x + rnorm(100L)
+  )
+  edf <- as_epi_df(tbl)
+
+  rec_tbl <- recipe(y ~ x, data = tbl)
+  rec_edf <- recipe(y ~ x, data = edf)
+  expect_snapshot(error = TRUE, epi_recipe(y ~ x, data = tbl))
+  erec_edf <- epi_recipe(y ~ x, data = edf)
+
+  ewf_rec_tbl <- epi_workflow(rec_tbl, linear_reg())
+  ewf_rec_edf <- epi_workflow(rec_edf, linear_reg())
+  ewf_erec_edf <- epi_workflow(erec_edf, linear_reg())
+
+  # above are all epi_workflows:
+
+  expect_s3_class(ewf_rec_tbl, "epi_workflow")
+  expect_s3_class(ewf_rec_edf, "epi_workflow")
+  expect_s3_class(ewf_erec_edf, "epi_workflow")
+
+  # but fitting drops the class or generates errors in many cases:
+
+  expect_s3_class(ewf_rec_tbl %>% fit(tbl), "epi_workflow")
+  expect_s3_class(ewf_rec_tbl %>% fit(edf), "epi_workflow")
+  expect_s3_class(ewf_rec_edf %>% fit(tbl), "epi_workflow")
+  expect_s3_class(ewf_rec_edf %>% fit(edf), "epi_workflow")
+  expect_snapshot(ewf_erec_edf %>% fit(tbl), error = TRUE)
+  expect_s3_class(ewf_erec_edf %>% fit(edf), "epi_workflow")
 })